Preliminary exploration of types
Up to this point, I'd taken the view that the MapState should be
extremely stateful, and that each MapStateUpdate would modify it without
leaving any trace of what was there before. Unfortunately,
in such a solution, it becomes impossible to implement
time-traveling in reverse performantly.

This was especially a concern with `Release`s, which could cover large
areas of the map, so reversing a release would result in recalculating
the proper CellStatus for every cell in that area.