Add events for mm-cell.
Added `'click'`, `'mouseover'`, and `'mouseout'` events for `mm-cell`.

Also fixed an issue with the raster metadata for the gripmap; I was
originally positioning cells relative to the top-left corner of the
origin cell, but I didn't think to update the metadata after I decided
to use the top-right corner instead.