Leaflet class to place divs on the map over cells
The idea behind this is to have `mm-map` cover the entire grid with
these hidden cells, and to use the cell's `mouseover` events to detect,
at the map level, when a user has hovered over a cell. The current
method to do so depends on `mousemove`, which does not update fast
enough to keep up with the user moving their mouse across many cells.

In the eventual system, `mm-map` would always keep track of the
currently hovered cell, and the appearance of individual cells or blobs
reacting would be created by descendant components reacting to the
current hoverCell, instead of each of them using Leaflet events
independently.

As an aside, the file at 'leaflet-classes/cell-div.ts' is very sloppily
written right now; I was not expecting to be able to get this working,
so I was writing in the style of low-investment experimentation. If this
invisible-cell-grid-on-top-of-blobs strategy yields satisfactory
performance, then I'll take the time to clean that up.