Further refine hoverCell performance
Unfortunately, I think I've reached the point of (more steeply)
diminishing returns; I tried a handful of different CSS `:hover` rules,
but they performed indistuingishably from simply useing the `mouseenter`
and `mouseleave` events on the `cellDiv`s themselves.

I did this with the intention of improving performance over the previous
method of rendering an mm-blob, and having it recalculate its coordinates
every time the mouse moved, but this made surprisingly little difference,
and the hoverCell UI is now hardcoded in `leaflet-classes/cell-grid.ts`.
Whether this hardcoding is a good thing or a bad thing is honestly difficult to
discern, but it can be reversed if it becomes an issue.