Remove square font
When I was attempting to improve the performance of `CellGrid`, I at one
point attempted to use characters in a square font instead of `div`s
with set height and width. It did not make a noticeable improvement, and
made it a bit more complicated to position the cells correctly, so this
idea was scrapped.