Reimplement `CellGrid` to fix Safari issue
The implementation of CellGrid, to this point, was a large div with
inline-block children of set height and width. This worked exactly as
intended in Firefox and Chrome, but in Safari, the rows of divs had a
very slight vertical overlap, and this caused cells to be misplaced as
you reached the bottom of the map. The overlap was about half of a
raster pixel between each row.

As I sought solutions to this problem, I could not find any tweaks that
could fix this, and it occurred to me after longer than it really
should have taken to use CSS grid directly, and this fixed the issue.

The resulting code is also much simpler (because it's no longer
necessary to individually resize each cell on zoom).