Initial WIP implementation of MapSnapshot logic
At the moment, I have not yet imported any claims, so I cannot fully
test this logic, and I have not tested any of it anyway. That said, I'm
quite satisfied by the logic in `utils/generate-claim-snapshot.ts`, and
I made a few small adjustments to the other types as well.